+++ changelog.md
@@ -4,6 +4,7 @@
 - Improve file structure
 - Move the Model Phpbb to PhpbbUser
 - Improve welcome email
+- Add the column province to the database groups
 
 ## v0.1.0
 - First version